refactor: tools.py split Phase 2 — extract email tools to _email.py
Moved IMAP/SMTP email tools (~399 lines) from _core.py into a dedicated
_email.py submodule.  Zero coupling to Playwright or ddl_checker — only
uses stdlib (imaplib, smtplib, email) + paths.

_core.py now ~4400 lines (down from ~4800 after Phase 1, ~5269 originally).
